PR18.6: address review feedback on trigger poller config
- Fix doc comments in TriggerPollerConfigSection to match the actual
  composition defaults (poll_interval_secs 30, fires_per_tick 32) so
  operators leaving fields unset get the rate they expect.
- Drop unused Serialize derive and the per-field
  #[serde(default, skip_serializing_if = "Option::is_none")] attrs on
  TriggerPollerConfigSection. The struct is Deserialize-only, so the
  serialize guards were dead; the field-level default is redundant
  because Option<T> already deserializes as None when absent. Matches
  every other *Section struct in the file.
- Reject [trigger_poller].fires_per_tick = 0 at config-parse time so
  operators see the error during boot config validation rather than
  later at spawn_trigger_poller. Symmetric with the existing
  poll_interval_secs > 0 and max_concurrent_fires_per_trigger == 1
  guards.
- Preserve the underlying parse error in the
  IRONCLAW_TRIGGER_POLLER_INTERVAL_SECS error message so values like
  "10s" or "1.5" tell the operator why they failed instead of just
  "must be a positive integer".
- Wrap test env-var mutation in a panic-safe RAII guard (restore on
  Drop) and serialize env-touching tests via a static LazyLock<Mutex>.
  Without this the new env tests race each other under the default
  parallel test harness; the previous manual snapshot/restore also
  leaked env state on assertion panic.
- Extract trigger_poller_settings + the env-mutation test harness into
  crates/ironclaw_reborn_cli/src/runtime/trigger_poller.rs. runtime.rs
  becomes runtime/mod.rs, dropping from 1231 to 883 lines (back below
  the 1000-line soft norm in .claude/rules/architecture.md). Keeps
  optional_nonempty_env in mod.rs as pub(super) since the child uses
  it.
- Clarify .env.example: IRONCLAW_TRIGGER_POLLER_ENABLED accepts 1/true
  to enable and 0/false to disable; any other value is a fatal
  startup error.
- Add four tests: fires_per_tick=0, IRONCLAW_TRIGGER_POLLER_ENABLED=yes
  (invalid), IRONCLAW_TRIGGER_POLLER_INTERVAL_SECS env override wins
  over config, and IRONCLAW_TRIGGER_POLLER_INTERVAL_SECS="10s"
  preserves the parse error context.
