+++ .env.example
@@ -212,8 +212,8 @@ HEARTBEAT_NOTIFY_USER=default
 # OFF by default. Enable to allow the host background poller to scan due trigger
 # records and auto-fire scheduled (cron) triggers. Without this, triggers can be
 # created and listed but will never fire automatically.
-# IRONCLAW_TRIGGER_POLLER_ENABLED=true           # set to true to opt in (default: false)
-# IRONCLAW_TRIGGER_POLLER_INTERVAL_SECS=60       # how often the poller checks for due triggers
+# IRONCLAW_TRIGGER_POLLER_ENABLED=true           # accepts 1/true to enable, 0/false to disable (default: false; any other value is a fatal startup error)
+# IRONCLAW_TRIGGER_POLLER_INTERVAL_SECS=60       # how often the poller checks for due triggers (positive integer seconds)
 
 # Docker Sandbox
 # SANDBOX_ENABLED=true
